requirements.txt
asgihandler==0.5.9
asgiref==3.5.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==40.0.2
daphne==4.0.0
Django==4.1.2
django-cors-headers==3.13.0
django-filter==2.4.0
djangorestframework==3.14.0
djangorestframework-csv==2.1.1
drf-spectacular==0.26.2
drf-spectacular-sidecar==2023.6.1
et-xmlfile==1.1.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
jsonschema==4.17.3
numpy==1.23.4
openpyxl==3.0.10
pandas==1.5.1
priority==1.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.1.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.5
PyYAML==6.0
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.2
tzdata==2022.5
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.12
zope.interface==6.0
whitenoise==6.6.0
pytest>=7.0
pytest-django>=4.5
pytest-cov>=4.0
factory-boy>=3.2
flake8>=6.0
celery>=5.3.0
redis>=5.0.0
django-allauth>=0.61.0
reportlab>=4.0
